requirements.txt
pyparsing==3.1.1
pySigma==0.10.6
PyYAML==6.0.1
sigma
Flask==3.0.0
Flask-Cors==4.0.0
